requirements.txt
requests>=2.31.0
urllib3>=2.0.0
PyYAML>=6.0
tqdm>=4.65.0
colorama>=0.4.6